pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "pics-server"
version = "0.1.0"
description = ""
authors = ["SweetBubaleXXX <70811772+SweetBubaleXXX@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
pydantic-settings = "^2.1.0"
dependency-injector = "^4.41.0"
sqlmodel = "^0.0.14"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {version = "^2.0.25"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.17"}
fastapi-jwt = "^0.2.0"
fastapi-pagination = "^0.12.14"
pydantic-extra-types = "^2.3.0"
aiofiles = "^23.2.1"
colorthief = "^0.2.1"
python-multipart = "^0.0.6"
fastapi-filter = {extras = ["sqlalchemy"], version = "^1.1.0"}
opencv-python = "^4.9.0.80"
numpy = "^1.26.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
factory-boy = "^3.3.0"
isort = "^5.13.2"
pytest-asyncio = "^0.23.3"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
